CTPA-291 cancellations. big uprn. action db fixes

+++ caseframesvc/src/main/java/uk/gov/ons/ctp/response/caseframe/domain/model/Case.java
@@ -34,7 +34,7 @@ public class Case implements Serializable {
   @Column(name = "caseid")
   private Integer caseId;
 
-  private Integer uprn;
+  private Long uprn;
 
   @Enumerated(EnumType.STRING)
   private CaseDTO.CaseState state;
